Week05 Observer (without steam)

+++ Minigin/Gamepad.cpp
@@ -18,12 +18,12 @@ public:
 
 	void UpdateInputImpl();
 
-	bool IsButtonUp(uint8_t button) const;
-	bool IsButtonDown(uint8_t button) const;
-	bool IsButtonPressed(uint8_t button) const;
+	bool IsButtonUp(uint16_t button) const;
+	bool IsButtonDown(uint16_t button) const;
+	bool IsButtonPressed(uint16_t button) const;
 
 	bool IsConnected() const;
-	static bool IsValidGamepadButton(uint8_t button);
+	static bool IsValidGamepadButton(uint16_t button);
 
 private:
 
@@ -72,17 +72,17 @@ void Gamepad::XinputImpl::UpdateInputImpl()
 	}
 }
 
-bool Gamepad::XinputImpl::IsButtonUp(uint8_t button) const
+bool Gamepad::XinputImpl::IsButtonUp(uint16_t button) const
 {
 	return m_ReleasedButtons & static_cast<WORD>(button);
 }
 
-bool Gamepad::XinputImpl::IsButtonDown(uint8_t button) const
+bool Gamepad::XinputImpl::IsButtonDown(uint16_t button) const
 {
 	return m_PressedButtons & static_cast<WORD>(button);
 }
 
-bool Gamepad::XinputImpl::IsButtonPressed(uint8_t button) const
+bool Gamepad::XinputImpl::IsButtonPressed(uint16_t button) const
 {
 	return m_Buttons & static_cast<WORD>(button);
 }
@@ -92,7 +92,7 @@ bool Gamepad::XinputImpl::IsConnected() const
 	return m_IsConnected;
 }
 
-bool Gamepad::XinputImpl::IsValidGamepadButton(uint8_t button)
+bool Gamepad::XinputImpl::IsValidGamepadButton(uint16_t button)
 {
 	constexpr WORD validButtons =
 		XINPUT_GAMEPAD_DPAD_UP | XINPUT_GAMEPAD_DPAD_DOWN |
@@ -124,17 +124,17 @@ void Gamepad::UpdateInput()
 	m_XinputImpl_Uptr->UpdateInputImpl();
 }
 
-bool Gamepad::IsButtonUp(uint8_t button) const
+bool Gamepad::IsButtonUp(uint16_t button) const
 {
 	return m_XinputImpl_Uptr->IsButtonUp(button);
 }
 
-bool Gamepad::IsButtonDown(uint8_t button) const
+bool Gamepad::IsButtonDown(uint16_t button) const
 {
 	return m_XinputImpl_Uptr->IsButtonDown(button);
 }
 
-bool Gamepad::IsButtonPressed(uint8_t button) const
+bool Gamepad::IsButtonPressed(uint16_t button) const
 {
 	return m_XinputImpl_Uptr->IsButtonPressed(button);
 }
@@ -144,7 +144,7 @@ bool Gamepad::IsConnected() const
 	return m_XinputImpl_Uptr->IsConnected();
 }
 
-bool Gamepad::IsValidGamepadButton(uint8_t button)
+bool Gamepad::IsValidGamepadButton(uint16_t button)
 {
 	return XinputImpl::IsValidGamepadButton(button);
 }
